Codenames: Handle case where one team hits the other's card causing them to win
One especially funny way for a game to end is when a player on one team selects the last card of the opposing team, causing them to win. Currently the logic for this doesn't work properly, as we only check whether a team won when that team plays a card of their own.

Fix this by checking whether the other team won if the current team accidentally plays one of their cards.

Also refactor a little bit to use existing helper methods and add some comments for convenience.